local-parser-server/requirements.txt
Flask==3.1.0
flask-cors==4.0.0
pdf2image==1.17.0
Pillow==10.4.0
pandas==2.2.2
openpyxl==3.1.5
openai>=1.0.0
python-dotenv>=1.0.0
xlrd>=2.0.1
